Package refactoring
- Remove coil from package name of common and compose sample module to avoid generating baseline profile rules for the sample module
