+++ coil-sample-common/src/main/java/sample/common/Application.kt
@@ -1,4 +1,4 @@
-package coil.sample
+package sample.common
 
 import android.app.Application
 import android.os.Build.VERSION.SDK_INT
@@ -10,7 +10,6 @@ import coil.decode.SvgDecoder
 import coil.decode.VideoFrameDecoder
 import coil.disk.DiskCache
 import coil.memory.MemoryCache
-import coil.sample.common.BuildConfig
 import coil.util.DebugLogger
 import okhttp3.Dispatcher
 import okhttp3.OkHttpClient
