+++ coil-sample-common/src/main/java/sample/common/Screen.kt
@@ -1,4 +1,4 @@
-package coil.sample
+package sample.common
 
 import coil.memory.MemoryCache
