+++ coil-sample-view/src/main/java/coil/sample/MainActivity.kt
@@ -18,6 +18,7 @@ import androidx.recyclerview.widget.StaggeredGridLayoutManager.VERTICAL
 import coil.load
 import coil.sample.databinding.ActivityMainBinding
 import kotlinx.coroutines.launch
+import sample.common.*
 
 class MainActivity : AppCompatActivity() {
